requirements.txt
fastapi~=0.115.12
pydantic~=2.11.5
sqlmodel~=0.0.24
python-dotenv~=1.1.0
SQLAlchemy~=2.0.41
uvicorn~=0.34.3
pytest~=8.4.0
httpx~=0.27.0
pytest-asyncio~=1.0.0